Added latest current code, not completed yet with changes
