+++ src/redux/guildsSlice.js
@@ -1,66 +1,71 @@
-// const { createSlice, createAsyncThunk, createEntityAdapter } = require('@reduxjs/toolkit');
+const { createSlice, createEntityAdapter, createAsyncThunk } = require('@reduxjs/toolkit');
 
-// const database = require('./database/database');
-// const ServerInfo = require('./database/models/dbdiscordserverinfo');
+// const database = require('../database/database');
+const ServerInfo = require('../database/models/dbdiscordserverinfo');
 
-// const guildsAdapter = createEntityAdapter();
+const guildsAdapter = createEntityAdapter();
 
-// export const selectAllGuilds = state => state.guildsAdapter
-// export const selectGuildById = (state, guildId) => state.guilds.find(guild => guild.id === guildId)
+const guildsSelector = guildsAdapter.getSelectors(state => state.guilds)
 
-// database.then(() => console.log("We're connecting to the DB from GuildsSlice")).catch(err => console.log(err));
+const fetchAllGuilds = createAsyncThunk(
+    'guilds/fetchAllGuilds', 
+    async (arg, thunkAPI) => {
+      const response = await ServerInfo.find({});
+      if(response) {
+        thunkAPI.dispatch(guildsSlice.actions.guildsLoaded(response));
+        return response;
+      } else {
+        return;
+      }
+    }
+);
 
-// const fetchAllGuilds = createAsyncThunk(
-//   'guilds/fetchAllGuilds', 
-//   async () => {
-//     const response = await ServerInfo.find({});
-//     console.log(response, "this is the response under guild/fetchAllGuilds");  
-//     return response;
-//   }
-// );
+const guildsSlice = createSlice({
+    name: 'guilds',
+    initialState: guildsAdapter.getInitialState(),
+    reducers: {
+      guildAdded(state, action) {
+        // console.log("we're here in guildAdded");
+        guildsAdapter.addOne(state, action.payload);
+      },
+      guildsLoaded(state, action) {
+        guildsAdapter.setAll(state, action.payload);
+      },
+      guildRemoved(state, action) {
+        guildsAdapter.removeOne(state, action.payload);
+      },
+      guildEatUpdated(state, action) {
+        const {id, value} = action.payload;
+        const guild = state.entities[id];
+        if(guild) {
+          guild.eat_updated = value;
+        }
+      },
+    },
+    extraReducers: {
+      [fetchAllGuilds.fulfilled]: (state, action) => {
+        console.log("The function fetchAllGuilds has been fulfilled");
+      },
+      // [fetchOneGuild.fulfilled]: (state, action) => {
+      //   console.log("*** This is the state in fetchOneGuilds.fulfilled ***\n", state);
+      //   console.log("*** This is the action in fetchOneGuilds.fulfilled ***\n", action);
+      //   // state.entities.push(action.payload)
+      // }
+    }
+});
 
-// const fetchOneGuild = createAsyncThunk(
-//   'guilds/fetchOneGuild', 
-//   async (guildId) => {
-//     const response = await ServerInfo.find({
-//       server_id: guildId
-//     });
-//     console.log(response, "this is the response under guild/fetchOneGuild")
-//     return response;
-//   }
-// )
+const {
+    guildAdded,
+    guildsLoaded,
+    guildRemoved,
+} = guildsSlice.actions
 
-// export const guildsSlice = createSlice({
-//   name: 'guilds',
-//   initialState: guildsAdapter.getInitialState(),
-//   reducers: {
-//     guildAdded: guildsAdapter.addOne,
-//     guildsLoaded(state, action) {
-//       guildsAdapter.setAll(state, action.payload);
-//     },
-//     guildEatUpdated(state, action) {
-//       const {id, value} = action.payload;
-//       const guild = state.entities[id];
-//       if(guild) {
-//         guild.eat_updated = value;
-//       }
-//     }
-//   },
-//   extraReducers: {
-//     [fetchAllGuilds.fulfilled]: (state, action) => {
-//       state.entities.push(action.payload);
-//     },
-//     [fetchOneGuild.fulfilled]: (state, action) => {
-//       state.entities.push(action.payload)
-//     }
-//   }
-// })
+module.exports = {
+    guildsSelector,
+    guildsSlice,
+    fetchAllGuilds,
+    guildAdded,
+    guildsLoaded,
+    guildRemoved
+}
 
-// export const { guildAdded, guildsLoaded, guildEatUpdated } = guildsSlice.actions;
-
-// export default guildsSlice.reducer;
-
-// export const {
-//   selectAll: selectAllGuilds,
-//   selectById: selectGuildById,
-// } = guildsAdapter.getSelectors(state => state.guilds)
